Add error reporting.

+++ build.sh
@@ -1,12 +1,33 @@
 #!/bin/bash
 
+MAINTAINER=pauls@ccs.neu.edu
+
+function error_check() {
+    if [ $? != 0 ] ; then 
+        mail -s "PRL website generation error" $MAINTAINER < errorlog
+        echo " -- [[[error]]] -- "
+        cat errorlog
+        echo " -- -- -- -- -- -- "
+        exit;
+    fi
+}
+
+mkdir -p output
+
 echo " -- gathering information from .prl directories -- "
-echo "cp will fail to find everything it wants in ~/.prl directories now.  This is normal."
 for u in `cat content/usernames`; do
-    cp /home/$u/.prl/bio.ss     content/people/$u.bio.ss
-    cp /home/$u/.prl/pubs.ss    content/pubs/$u.pubs.ss
-    cp /home/$u/.prl/pubs.bib   content/pubs/$u.pubs.bib
-    cp -r /home/$u/.prl/static  output/static-$u
+    if [ -e /home/$u/.prl/bio.ss ] ; then
+        cp /home/$u/.prl/bio.ss     content/people/$u.bio.ss
+    fi
+    if [ -e /home/$u/.prl/pubs.ss ] ; then
+        cp /home/$u/.prl/pubs.ss    content/pubs/$u.pubs.ss
+    fi
+    if [ -e /home/$u/.prl/pubs.bib ] ; then
+        cp /home/$u/.prl/pubs.bib   content/pubs/$u.pubs.bib
+    fi
+    if [ -d /home/$u/.prl/static ] ; then
+        cp -r /home/$u/.prl/static  output/static-$u
+    fi
 done
 
 cp content/centralized-people/* content/people/
@@ -24,9 +45,12 @@ cp content/pubs/*.xml working/
 cp content/pubs/*.ss  working/
 
 cd working/
-mzscheme ../xml-to-sxml.ss *.xml
+mzscheme ../xml-to-sxml.ss *.xml 2> errorlog
+error_check
 
-bibtex allpubs
+bibtex allpubs 2> errorlog
+error_check
+
 mv allpubs.bbl bibtexpubs.withbackslashes
 
 #the backslashes get us into trouble otherwise
@@ -36,6 +60,7 @@ cd ../
 
 echo " -- generating HTML -- "
 
-mzscheme render-html.ss
+mzscheme render-html.ss 2> errorlog
+error_check
 
 echo " -- done -- "
